Added working directory to installer call bcs it doens't install otherwise (and other minor changes)

+++ RiverbirdInstaller.ps1
@@ -228,46 +228,60 @@ try {
         $ExitCode = 0
     } else {
         Log "Riverbird agent isn't installed already. Proceeding to install..."
-        <# 
-            Create directory for installation file
-        #>
 
-        Log "Checking if destination directory '$( $DirDest )' exists." 
-        if (-not ( Test-Path -Path $DirDest )) {
-            Log "Doesn't exist. Creating..."
-            New-Item -Path $DirDest -ItemType Directory -Force
+        Log "Checking if installer '$( $FullPathInstaller )' exists." 
+        if (-not ( Test-Path -Path $FullPathInstaller )) {
+            Log "Doesn't exist. Proceeding..."
+
+            <# 
+                Create directory for installation file
+            #>
+
+            Log "Checking if destination directory '$( $DirDest )' exists." 
+            if (-not ( Test-Path -Path $DirDest )) {
+                Log "Doesn't exist. Creating..."
+                New-Item -Path $DirDest -ItemType Directory -Force
+            } else {
+                Log "Exist already."
+            }
+            
+            <# 
+                Connect to FTP server to receive installation file
+            #>
+
+            # Build credentials for FTP server
+            $SecureString = ConvertTo-SecureString -AsPlainText $FtpPassword -Force
+            $Creds = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList $FtpUsername, $SecureString 
+
+            try {
+                # Connect to FTP server
+                $s = New-SFTPSession -ComputerName $FtpServerFqdn -Credential $Creds -Port 22 -AcceptKey:$true
+
+                # Download installation file
+                Get-SFTPItem -SFTPSession $s -Path $FullPathSrc -Destination $DirDest -Force # One can only specify a directory as destination. The file will always keep its name.
+            }
+            finally {
+                # Disconnect from FTP server
+                Remove-SFTPSession -SFTPSession $s
+            }
         } else {
             Log "Exist already."
-        }
-
-        <# 
-            Connect to FTP server to receive installation file
-        #>
-
-        # Build credentials for FTP server
-        $SecureString = ConvertTo-SecureString -AsPlainText $FtpPassword -Force
-        $Creds = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList $FtpUsername, $SecureString 
-
-        try {
-            # Connect to FTP server
-            $s = New-SFTPSession -ComputerName $FtpServerFqdn -Credential $Creds -Port 22 -AcceptKey:$true
-
-            # Download installation file
-            Get-SFTPItem -SFTPSession $s -Path $FullPathSrc -Destination $DirDest -Force # One can only specify a directory as destination. The file will always keep its name.
-        }
-        finally {
-            # Disconnect from FTP server
-            Remove-SFTPSession -SFTPSession $s
         }
 
         <# 
             Start installation
         #>
 
-        Start-Process -FilePath $FullPathInstaller -ArgumentList $Arguments
-
-        Log 'Started installation successfully. Exiting successfully...'
-        $ExitCode = 0
+        Log 'Starting installation...'
+        $Process = Start-Process -FilePath $FullPathInstaller -WorkingDirectory $DirDest -ArgumentList $Arguments -Wait -PassThru
+        
+        if ($Process.ExitCode -eq 0) {
+            Log 'Installation Exit Code was 0, thus installation was successful. Exiting successfully.'
+            $ExitCode = 0
+        } else {
+            Log 'Installation Exit Code was not 0, thus installation was unsuccessful. Exiting unsuccessfully.'
+            $ExitCode = 1
+        }
     }
 
     #endregion EXECUTION
